refactor: extract Form_Records_Files from Form_Submissions
Move uploaded-file cleanup (delete on record delete + path helpers) into a
dedicated Form_Records_Files class that registers its own before_delete_post
hook. Behaviour unchanged; covered by the existing permanent-delete tests.
